+++ CHANGELOG.md
@@ -2,6 +2,10 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+### Enhancements
+
+* Optimized server-side processing of submitted programs.
+
 ## v21.9.21
 
 ### Bugfixes